+++ fasttuple-bench/pom.xml
@@ -26,11 +26,11 @@ questions.
     <modelVersion>4.0.0</modelVersion>
     <parent>
     <artifactId>fasttuple</artifactId>
-    <groupId>com.boundary</groupId>
+    <groupId>com.nickrobison</groupId>
     <version>0.1-SNAPSHOT</version>
   </parent>
 
-    <groupId>com.boundary</groupId>
+    <groupId>com.nickrobison</groupId>
     <artifactId>fasttuple-bench</artifactId>
     <version>0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
@@ -50,7 +50,7 @@ questions.
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>com.boundary</groupId>
+            <groupId>com.nickrobison</groupId>
             <artifactId>fasttuple-core</artifactId>
             <version>${project.version}</version>
         </dependency>
