--- fasttuple-bench/src/main/java/com/boundary/tuple/Container.java
@@ -1,16 +1,0 @@
-package com.boundary.tuple;
-
-/**
-* Created by cliff on 5/12/14.
-*/
-class Container {
-    public long a;
-    public int b;
-    public short c;
-
-    public Container(long a, int b, short c) {
-        this.a = a;
-        this.b = b;
-        this.c = c;
-    }
-}
