+++ fasttuple-bench/src/main/java/com/nickrobison/tuple/Container.java
@@ -1,0 +1,16 @@
+package com.nickrobison.tuple;
+
+/**
+* Created by cliff on 5/12/14.
+*/
+class Container {
+    public long a;
+    public int b;
+    public short c;
+
+    public Container(long a, int b, short c) {
+        this.a = a;
+        this.b = b;
+        this.c = c;
+    }
+}
