+++ fasttuple-core/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.boundary</groupId>
+    <groupId>com.nickrobison</groupId>
     <artifactId>fasttuple-core</artifactId>
     <version>0.1-SNAPSHOT</version>
