--- fasttuple-core/src/main/java/com/boundary/tuple/Destroyer.java
@@ -1,8 +1,0 @@
-package com.boundary.tuple;
-
-/**
- * Created by cliff on 5/15/14.
- */
-public interface Destroyer<T> {
-    public void destroyArray(T[] ary);
-}
