--- fasttuple-core/src/main/java/com/boundary/tuple/Loader.java
@@ -1,8 +1,0 @@
-package com.boundary.tuple;
-
-/**
- * Created by cliff on 5/15/14.
- */
-public interface Loader<T> {
-    public T[] createArray(int size) throws Exception;
-}
