+++ fasttuple-core/src/main/java/com/nickrobison/tuple/Destroyer.java
@@ -1,0 +1,8 @@
+package com.nickrobison.tuple;
+
+/**
+ * Created by cliff on 5/15/14.
+ */
+public interface Destroyer<T> {
+    public void destroyArray(T[] ary);
+}
