+++ fasttuple-core/src/main/java/com/nickrobison/tuple/Loader.java
@@ -1,0 +1,8 @@
+package com.nickrobison.tuple;
+
+/**
+ * Created by cliff on 5/15/14.
+ */
+public interface Loader<T> {
+    public T[] createArray(int size) throws Exception;
+}
